shader hot-reloading
This is an initial version that is functional, but requires cleanup.

The macro now returns a `ShaderData` instead of `'static &[u32]`. Calling the `data` method on the shader data will return an `impl Deref<Target = [u32]>`, which will contain the newest shader data.

There is a `should_recompile` function at the crate root that can be used to query for *any* shader changes.

Internally, a file watcher is used to detect any changes to relevant files. This sets a global flag that shaders should be recompiled.

A shader will only get recompiled if its file modify time differs from the times that were used for the pre-computed shader.

Exporting the `ShaderData` struct and `should_recompile` function required splitting into two crates, because a proc-macro crate cannot export normal items.